Rename module that serialize request to prepare response serialization
