+++ src/blacksmith/__init__.py
@@ -72,7 +72,7 @@ from .service._async.route_proxy import AsyncRouteProxy
 from .service._sync.base import SyncAbstractTransport
 from .service._sync.client import SyncClient, SyncClientFactory
 from .service._sync.route_proxy import SyncRouteProxy
-from .service.request_serializer import (
+from .service.http_body_serializer import (
     AbstractRequestBodySerializer,
     register_request_body_serializer,
     unregister_request_body_serializer,
